Add initiate requisition test case for controller
This is only for the "happy path" test case. Also add mocking methods for reference data classes, and a initiate requisition DTO.

The previous initiate requisition logic does not quite make sense, as there is no requisition to pass in the request body for initiation. Rather, the body should pass in values for facility, program, processing period and emergency requisition, then return an empty, initiated requisition. This is also true for the service logic, so the service logic has been changed to use facility, program, processing period, rather than a requisition object.