Added importer/exporter pattern to Requisition and RequisitioDto
Now /requisition/id PUT endpoint requires RequisitionDto as body.
Then it imports this RequisitionDto to Requisition and updates it in database.
This is work in progress - there are two static constructors needed -
the places where the code is not completed are marked by TODO comments (Two comments in Requisition domain).